[PROF-12841] Fix profiler not identifying executables with gems they belong to
**What does this PR do?**

This PR fixes the profiler's "code provenance" metadata to include
the paths to gem executables (if any).

This ensures that, when e.g. starting `puma` via `bin/puma` or
`bundle exec puma`, we still identify the puma executable as
belonging to the `puma` gem.

**Motivation:**

The "code provenance" is used to power the "Only My Code" feature
that shows up in multiple places in the profiler UX, and so we want
it to be as accurate as possible.

**Additional Notes:**

N/A

**How to test the change?**

This change includes test coverage. It can also be tested easily
by running

```bash
DD_PROFILING_ENABLED=true DD_SERVICE=test-provenance bundle exec ddprofrb exec pry -e "sleep 1; exit"
```

and checking that `bin/pry` is correctly categorized as belonging
to the `pry` gem in the Datadog UX.