[NO-TICKET] Add `@DataDog/ruby-guild` as fallback codeowner on rbs folders
**What does this PR do?**

This PR adds `@DataDog/ruby-guild` as an additional codeowner on
rbs folders.

**Motivation:**

While looking at @vpellan's PR
https://github.com/DataDog/dd-trace-rb/pull/5014 it seems overkill to
need reviews from across n teams for type signature changes.

I think at this point we want to encourage all rbs contributions and
lower their overhead, not make it really annoying to get them into the
project, so the more reviewers the better.

**Additional Notes:**

N/A

**How to test the change?**

CODEOWNERS changes are always annoying to test; on the other hand, if
anything is wrong, it's easy to revert.