[PROF-12941] Upgrade libdatadog dependency to version 23.0.0
**What does this PR do?**

This PR bumps the libdatadog dependency from version 22.0.1.1.0 to
version 23.0.0.1.0.

As this new libdatadog version now compresses profiles using zstd
instead of lz4, I've replaced one with the other on our tests.

**Motivation:**

Adopt the new, better compression for the Ruby profiler.

**Additional Notes:**

Other than the switch from lz4 to zstd on the Ruby profiler, I did
not need to change anything to get a green test suite locally.

**How to test the change?**

I've tested this locally as I'm preparing to release the libdatadog
gem change. I'm going to open this PR as a draft until the libdatadog
gem is up on rubygems.