[PROF-13286] Disable heap profiling on Ruby 4 due to incompatibility
**What does this PR do?**

This PR disables the heap profiling feature when running on Ruby 4 as
the current implementation (that relies on object_id) is not safe to use
on Ruby 4.

**Motivation:**

Avoid impacting applications running on Ruby 4.

**Additional Notes:**

Discussion with the Ruby core team in
https://bugs.ruby-lang.org/issues/21710 has uncovered that our use
of object_id from inside the NEWOBJ tracepoint is not safe on
Ruby 4, and it's likely it can't be made safe.

Until we're able to replace the Datadog profiler's usage of
object_id for heap profiling, we won't be able to safely offer
this feature on Ruby 4.

(We temporarily disabled the heap profiler for 4.0.0-preview2 in
 #5091 and re-enabled it in #5121 when it looked fixed but
discussion on the upstream ticket uncovered further issues)

**How to test the change?**

Validate that profiler doesn't enable heap profiling on Ruby 4.