[PROF-13115] Disable heap profiling on Ruby 4 preview due to incompatibility
**What does this PR do?**

This PR disables the heap profiling feature when running on Ruby 4 as
right now it can trigger the crash reported in
https://bugs.ruby-lang.org/issues/21710 .

**Motivation:**

Until either https://bugs.ruby-lang.org/issues/21710 gets fixed or
we change the heap profiler to use the new
[`rb_gc_mark_weak` API](https://bugs.ruby-lang.org/issues/19783)
it's not safe to enable the heap profiling feature on Ruby 4.

**Additional Notes:**

I'm preparing a PR with other, smaller fixes for the profiler
Ruby 4 support, but I decided to split this one by itself so we
could have a separate, clear, changelog entry about this change,
as well as a single PR to discuss this subject.

**How to test the change?**

We don't yet have Ruby 4 in CI, so this can only be tested locally or
by mocking the `RUBY_VERSION`.