[PROF-13115] Bootstrap installing dependencies on Ruby 4.0.0-preview2
**What does this PR do?**

This PR adds the `ruby-4.0.gemfile` file that's needed for running
`bundle install` on Ruby 4.0.0-preview2.

The contents of this file being added are copy/pasted from
`ruby-4.0.gemfile` with no changes.

I did not add or change anything on purpose: This is a very minimal
part of the Ruby 4 CI support being prepared in
https://github.com/DataDog/dd-trace-rb/pull/5084 but
unlocks local development with Ruby 4.0. I suggest we keep
adjustments/cleanups/etc for that later PR where we officially add Ruby
4.0 to our CI.

**Motivation:**

Without this file it's impossible to do local development on
Ruby 4.0.0-preview2.

It also came up in
<https://github.com/DataDog/datadog-ruby_core_source/pull/19>:
we want to check if the profiler extension compiles on Ruby 4.0 and
we're not able without this file.

With this PR we can run `bundle install` and start running tests
and tackling what's missing.

**Additional Notes:**

N/A

**How to test the change?**

Validate that it's possible to run `bundle install` on Ruby 4.0.