[PROF-13115] Fix profiler ractor specs failing on Ruby 4
**What does this PR do?**

This PR updates the profiler specs to handle the removal of
[Ractor#take in Ruby 4.0](https://github.com/ruby/ruby/commit/57d10c6e806c4572eb75084462ecdf883591f4ea)

To keep the specs working on 3.x Rubies, we still use Ractor#take there.

**Motivation:**

Get a green CI with profiling on Ruby 4.0.0-preview2.

**Additional Notes:**

Thanks @lloeki for flagging this one -- I had missed it on my previous
"support 4.0.0-preview2 PRs" as the ractor specs only run when
`-t ractors` gets passed in (and CI does that as well).

**How to test the change?**

Run these specs with `-t ractors` and observe they're green on Ruby 3
and 4.