Fix Karafka framework patch
By the time that `Karafka::App#initialized!` is called,
`Karafka.producer` was already initialized.

On top of that, there's really nowhere we can hook in the Karafka
initialization where we can be sure that `Karafka.producer` wasn't yet
initiaized - that's because the Karafka initialization is NOT
necessarily tied to the WaterDrop initialization. For instance, the
following is a possible scenario:

```
$producer = WaterDrop::Producer.new { ... }

Datadog.configure do |c|
  c.tracing.instrument :karafka
end

 # note that the producer was initialized before the Karafka app (and in
 # this case, even before datadog was configured)
Karafka::App.setup do |c|
  c.producer = $producer
end
```

So instead of trying to hook somewhere before `Karafka.producer` is
initialized, let's simply listen to a Karafka after-initialization event
and append our middleware to the producer when that event is triggered.