[PROF-13273] Fix profiler causing `Bundler::PermissionError`
**What does this PR do?**

This PR fixes a regression introduced in #4999: In that PR we added a
call to `Bundler.bin_path` but did not realize that bundler not only
returns the path **BUT CRUCIALLY** it makes sure the path exists
by going as far as trying to create it if it doesn't.

If the directory does not exist, and the application directory is
read-only (or the app does not have enough permissions), the
application will fail with:

> There was an error while trying to write to (directory). It is likely
> that you need to grant write permissions for that path.
> (Bundler::PermissionError)

Fixes #5137

**Motivation:**

Fix a regression introduced in #4999 that shipped in v2.23.0.

**Additional Notes:**

The bundler's `bin_path` method has had a similar
behavior/implementation for 15+ years:

* master -> https://github.com/ruby/rubygems/blob/570c3419c7f111d8665710f4e2cb15ca878e606d/bundler/lib/bundler.rb#L119-L126
* Jul 10, 2010 -> https://github.com/ruby/rubygems/commit/f799489451b9b50106ecb9fd5890bebcd47ccde6

So we should be fine in replicating its behavior + having a
test that checks we're matching it.

Worst case, we get `CodeProvenance` slightly off, which
compared to "broke the app", seems better.

I've also added a `rescue` to the method so hopefully
we don't run into more surprises here.

(Note that while we could've kept calling `Bundler.bin_path`,
wrapping it with the `rescue`, this would still produce the
side-effect of creating the folder and it seems weird to
me to end up in this situation of "turning on the profiler
causes bin/ to be created". This is why I chose to reimplement
the behavior, minus folder creation)

**How to test the change?**

I've updated the test coverage.

Furthermore, the repro mentioned in #5137 also works to
show it's fixed:

```bash
$ rm -rf bin/
$ chmod -w .
$ bundle exec rake compile && DD_PROFILING_ENABLED=true bundle exec ruby -rdatadog/profiling/preload -e "puts 'hello'"
```