Assert on empty standard error in execution spec. (#5149)
The idea here is that if the test fails, there usually would be
some error output on standard error and standard output may be
empty. Thus having an assertion on standard output containing
expected text fails the test, but does not help with troubleshooting.

The extra assertion on standard error being empty is meant
to reveal the actual issue.

Co-authored-by: Oleg Pudeyev <code@olegp.name>